delete duplicate parsing file
